+++ arduino-main/imu.h
@@ -2,12 +2,26 @@
 #define IMU_H
 
 #include <Arduino.h>
+#include "input.h"
+#include "communication.h"
+
 #include <Adafruit_Sensor.h>
 #include <Adafruit_BNO055.h>
 #include <utility/imumaths.h>
 
-IMU(int inputPin, String incomingPartID);
+/*
+  The IMU class reads data from the IMU's sensors (position, acceleration, temperature) and sends this to the Pi using the communication class.
+*/
+class IMU: public Input {
 
-int getValue();
+  protected:
+    bool initialised;
+    Adafruit_BNO055 imu;
+
+  public:
+    IMU(int inputPin, String incomingPartID);
+
+    int getValue();
+};
 
 #endif
